fix: correct State.sequence type to float and WorkItem.state to accept expanded objects
- State.sequence, CreateState.sequence, UpdateState.sequence: int -> float
  The Plane API uses a Django FloatField for state sequence values (e.g.
  -57052.5, 8482.5), but the model declared it as int, causing Pydantic
  validation errors.

- WorkItem.state: str -> str | StateLite
  When the expand query parameter includes 'state', the API returns an
  expanded state object instead of a UUID string. WorkItemDetail already
  handled this correctly, but WorkItem (used by PaginatedWorkItemResponse)
  did not.

Fixes makeplane/plane-mcp-server#83

+++ plane/models/states.py
@@ -16,7 +16,7 @@ class State(BaseModel):
     name: str
     description: str | None = None
     color: str
-    sequence: int | None = None
+    sequence: float | None = None
     group: GroupEnum | None = None
     is_triage: bool | None = None
     default: bool | None = None
@@ -47,7 +47,7 @@ class CreateState(BaseModel):
     name: str
     description: str | None = None
     color: str
-    sequence: int | None = None
+    sequence: float | None = None
     group: GroupEnum | None = None
     is_triage: bool | None = None
     default: bool | None = None
@@ -63,7 +63,7 @@ class UpdateState(BaseModel):
     name: str | None = None
     description: str | None = None
     color: str | None = None
-    sequence: int | None = None
+    sequence: float | None = None
     group: GroupEnum | None = None
     is_triage: bool | None = None
     default: bool | None = None
